+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## WIP
+
+- Fix panics in `renet2_setup::ClientCounts` methods.
+
 ## 0.3.0 - 02/07/25
 
 - Replace `renet2_setup` crate's `ws_certs` feature with `ws-native-tls` and `ws-rustls` features.